Use std::ceil to relax max_differences_with_pattern

+++ include/fuzzy/agenda_item.hh
@@ -17,7 +17,7 @@ namespace fuzzy
     unsigned s_id;
     std::vector<bool> map_pattern; /* set of unigrams that are both in the pattern and in the suffix
                                       map_pattern[i]==true means that the word pattern[i] appears in the suffix */
-    int coverage; // number of true values in map_pattern
+    size_t coverage; // number of true values in map_pattern
     size_t maxmatch; // maximum length of the longest subsequence between the sentence associated with the AgendaItem and the pattern
   };
 }
